Completer BoxPlotDrawer and improved FileExporter.

+++ src/de/etrayed/boxplotgen/plot/BoxPlotValueCalculator.java
@@ -7,7 +7,7 @@ import java.util.Arrays;
  */
 class BoxPlotValueCalculator {
 
-    static BoxPlotInfo withValues(double[] values, float scaling) {
+    static BoxPlotInfo withValues(double[] values, int scaling) {
         double minimum;
         double lowerQuartile;
         double median;
